fix: resolve PR review issues — noIndex, response extraction, useAIChatSafe, slugify utils
- Remove noIndex: true from docusaurus.config.js (was de-indexing entire site)
- Fix CodeExamples response panel to extract .example from response wrapper object
- Replace useAIChat() with useAIChatSafe() in CodeExamples to prevent crash outside provider
- Extract slugify to src/component/ApiReference/utils.js (shared utility, removes cross-layer import)
- Update ApiSidebar and _ApiDocPage to import slugify from utils
